Código refatorado. Novas pastas: Models, Repository, Service, Util e View.
