+++ src/Models/Cliente.java
@@ -1,3 +1,5 @@
+package Models;
+
 public class Cliente {
     String nome;
     String CPF;
