+++ src/Models/Nota.java
@@ -1,3 +1,5 @@
+package Models;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
