+++ test/testCliente.java
@@ -1,3 +1,4 @@
+import Models.Cliente;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
