+++ test/testProduto.java
@@ -1,3 +1,4 @@
+import Models.Produto;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
